requirements.txt
git+ssh://git@github.com:Mmesek/mdiscord.git
git+ssh://git@github.com:Mmesek/mlib.git
sqlalchemy
psycopg2-binary
redis
influxdb-client
python-git
python-i18n